refactor: create file for transform responses; rewrite isLoading in Movie page
