+++ src/services/moviesService.ts
@@ -1,5 +1,10 @@
 import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/dist/query/react';
 
+import {
+	transformResponseFetchAllMovies,
+	transformResponseFetchMovieById,
+} from './transformResponses/moviesTransfromResponse';
+
 import type { MovieByIdResponse, MoviesResponse } from 'types/movies';
 
 export const moviesAPI = createApi({
@@ -16,62 +21,15 @@ export const moviesAPI = createApi({
 			query: (params: { page: number; limit: number }) => ({
 				url: `movie?page=${params.page}&limit=${params.limit}`,
 			}),
-			transformResponse: (responseData: MoviesResponse) => {
-				const { total, page, pages, limit } = responseData;
-
-				const data = responseData.docs.map((item) => {
-					return { ...item, previewUrl: item.poster.previewUrl };
-				});
-
-				return {
-					total,
-					page,
-					pages,
-					limit,
-					data,
-				};
-			},
+			transformResponse: (responseData: MoviesResponse) =>
+				transformResponseFetchAllMovies(responseData),
 		}),
 		fetchMovieById: build.query({
 			query: (params: { movieId: number }) => ({
 				url: `movie/${params.movieId}`,
 			}),
-			transformResponse: (responseData: MovieByIdResponse) => {
-				const {
-					poster,
-					genres,
-					rating,
-					name,
-					alternativeName,
-					year,
-					description,
-					ageRating,
-					countries,
-					fees,
-					watchability,
-				} = responseData;
-
-				return {
-					url: poster.url,
-					genres,
-					rating,
-					name,
-					alternativeName,
-					year,
-					description,
-					ageRating,
-					countries,
-					feesRussia: fees.russia,
-					feesWorld: fees.world,
-					watchability: watchability.items.map((item) => {
-						return {
-							linkIcon: item.logo.url,
-							name: item.name,
-							url: item.url,
-						};
-					}),
-				};
-			},
+			transformResponse: (responseData: MovieByIdResponse) =>
+				transformResponseFetchMovieById(responseData),
 		}),
 	}),
 });
